made widgets/layouts for score keeping app
splash screen - where progress bar , app name and developer name is displayed. flow layout is used in xml , comments are present.

main activity - where user has the option to select game type with spinner and also user can select countries of two teams from spinners .

image of image is displayed in spinner with game name.

when user chnge the countries ther flags are updatd in image view.

all the widgets will work in next commit.

buttons to increase and decrease score are present but not working .
as it is just a layout , next commit will include functionality.

radio button to change score points are included and also there is a check box at top to stop or run game music.

switch is there to change current scoring team.there are two text views to show current score of both the teams  when increased or decreased.